+++ guest/commands/wifirouter/router.cc
@@ -171,8 +171,7 @@ void RouteWIFIPacket(nl_sock* nl, int simfamily, ClientsTable* clients,
     LOG(INFO) << "Received netlink packet from " << std::hex << key;
     for (auto it = targets->find(key); it != targets->end() && it->first == key;
          ++it) {
-      auto num_written =
-          send(it->second, hdr, hdr->nlmsg_len, MSG_NOSIGNAL);
+      auto num_written = send(it->second, hdr, hdr->nlmsg_len, MSG_NOSIGNAL);
       if (num_written != static_cast<int64_t>(hdr->nlmsg_len)) {
         pending_removals.insert(it->second);
       }
@@ -188,10 +187,11 @@ bool HandleClientMessage(int client, MacToClientsTable* targets) {
   std::unique_ptr<nlmsghdr, void (*)(nlmsghdr*)> msg(
       reinterpret_cast<nlmsghdr*>(malloc(kMaxSupportedPacketSize)),
       [](nlmsghdr* h) { free(h); });
-  int64_t size = recv(client, msg.get(), kMaxSupportedPacketSize, 0);
+  ssize_t size = recv(client, msg.get(), kMaxSupportedPacketSize, 0);
 
   // Invalid message or no data -> client invalid or disconnected.
-  if (size == 0 || size != msg->nlmsg_len || size < sizeof(nlmsghdr)) {
+  if (size == 0 || size != static_cast<ssize_t>(msg->nlmsg_len) ||
+      size < static_cast<ssize_t>(sizeof(nlmsghdr))) {
     return false;
   }
